Keep map component reusable
